Reimplement handling of the changes feed request
the $Couch.changes function now returns an object by which you can
start and stop the feed, and subscribe to its `on_change` and `on_error`
events.

The implementation still reconnects if needed, on error and when it
receives new data (the connection is closed then). You must explicitelly
stop it to make it stop!

I used jQuerys events to signal new data and errors (bind and trigger).