Simplify the watchdog
The watchdog in the _changes handler was made more stupid but also robust.
Previously it would "ping" the database with http requests and abort the long-poll
request if there was no answer.

But that didn't solve the case when a short interuption would loose the FIN or
RST TCP packet for the long-poll connection. The long-poll request would be dead
but the ping would still succeed.

Now, I just blindly abort the long-poll after some time has passed (heartbeat * 10).
The worst case is it wont know it's offline for that time.

It would be much better to check the long-poll request if it received any new data,
but that's not supported in all browsers (works in Firefox though).