add DoInitialView command
The init function no longer calls the Couch.getLast100Messages function directly,
instead it invokes a DoInitialView command which calls the function.
That way, we can retry the request if it fails without duplicating the
code.

also move updates to the model in separate functions to remove some
indentation.

and update some more elm style