fix a subtle (aysnc) bug when loading backhistory
if you press the button several times before the reponse from the server
comes, it'll start several requests which are doing the same thing
(based on the first element of the channelLog.messages).

so carry the head of the channelLog.messages with the response and only
update the channelLog if it matches the current head.