fix a bug with pagination
some checks were needed to limit the async response to only take the
result if the model didn't change (or that part of the model).

but you can't pattern match against variables in elm (like in
erlang), so instead it just overwrote those variables. change to `if`
with some helper functions.